refactor(handler): 抽取 gateway_handler TK companion，收薄 Messages 热点
把 protocol execute、Kiro 非 failover、failover enrich、Usage errgroup、
universal/antigravity models 迁到 *_tk_*.go；同步 SSOT/sentinel 锚定到 companion。
功能不变。no-web-impact
