+++ scripts/sentinels/gateway-tk.json
@@ -2182,7 +2182,7 @@
       "rationale": "OAuth Claude Code mimicry beta registry: Sonnet/Opus and Haiku sets must stay aligned with real cc CLI traffic (latest cc capture; version single-sourced in deploy/aws/stage0/anthropic-http-mimicry-baselines.json). Dropping thinking-token-count / structured-outputs or reverting Haiku to the legacy claude-code + extended-cache-ttl set re-opens third-party cohort signals on OAuth subscriptions."
     },
     {
-      "path": "backend/internal/handler/gateway_handler.go",
+      "path": "backend/internal/handler/gateway_handler_tk_kiro.go",
       "must_contain": [
         "var canonicalUARejectErr *service.CanonicalIngressUARejectedError",
         "errors.As(err, &canonicalUARejectErr)",
@@ -4087,9 +4087,8 @@
       "rationale": "Focused regressions proving the issue #625 boundary: a status-0 `context canceled` transport error is client-owned; deadline/timeout remains provider-owned; a real final upstream 5xx remains provider-owned; an older failover 5xx followed by terminal cancel is client-owned; the reverse ordering remains provider-owned; and a final 499 remains client-owned even with earlier upstream events. Dropping these lets an upstream merge silently revert the quad-P0 fix or reintroduce failover-history pollution."
     },
     {
-      "path": "backend/internal/handler/gateway_handler.go",
+      "path": "backend/internal/handler/gateway_handler_tk_forward_error.go",
       "must_contain": [
-        "func (h *GatewayHandler) ensureForwardErrorResponse(c *gin.Context, streamStarted bool) bool",
         "func (h *GatewayHandler) ensureForwardErrorResponseForError(c *gin.Context, err error, streamStarted bool) bool",
         "if isClientClosedRequest(c, err) {",
         "markClientClosedForwardRequest(c)"
@@ -6237,9 +6236,9 @@
       "rationale": "TK channel pricing restriction companion: pre-scheduling model restriction checks based on channel-level pricing configuration. Extracted from gateway_service.go to reduce upstream merge surface. Forward and ForwardCountTokens call checkChannelPricingRestriction; dropping the companion breaks compilation."
     },
     {
-      "path": "backend/internal/handler/gateway_handler.go",
+      "path": "backend/internal/handler/gateway_handler_tk_usage.go",
       "must_contain": [
-        "g.Go(func() error { usageData = h.buildUsageData(gctx, apiKey.ID); return nil })"
+        "g.Go(func() error { usageData = h.buildUsageData(gctx, apiKeyID); return nil })"
       ],
       "rationale": "GatewayHandler.Usage (/key-usage) perf fix: three query groups must run concurrently via errgroup. Reverting to sequential best-effort blocks restores prod TTFB with identical response bytes. Sibling perf-query-shape sentinel pins the same shape; this entry satisfies the gateway-tk registry update gate when the upstream-shaped handler is edited with deletions."
     },
@@ -6617,9 +6616,9 @@
       "rationale": "A classified internal relay empty-pool response must preserve request-local exclusions at terminal selection exhaustion; clearing them retries the same relay and can manufacture the three-hit cooldown inside one client request."
     },
     {
-      "path": "backend/internal/handler/gateway_handler.go",
+      "path": "backend/internal/handler/gateway_handler_tk_forward_error.go",
       "must_contain": [
-        "failoverErr.ClientStatusCode > 0",
+        "failoverErr.ClientStatusCode <= 0",
         "failoverErr.ResponseHeaders.Get(\"Retry-After\")",
         "failoverErr.ClientMessage"
       ],
@@ -7152,7 +7151,7 @@
       "rationale": "Codex discovery may reuse an upstream manifest only as a schema template. Automatic keys must return a native empty manifest for a genuine no-capability set, otherwise filter models[] to capability-authorized slugs while preserving unknown manifest fields; removing either projection turns an empty business set into 503 or advertises every upstream model."
     },
     {
-      "path": "backend/internal/handler/gateway_handler.go",
+      "path": "backend/internal/handler/gateway_handler_tk_model_list.go",
       "must_contain": [
         "requestLogger(c, \"gateway.models\", zap.String(\"protocol\", string(protocol))).Warn(\"capability_discovery_failed\", zap.Error(err))",
         "requestLogger(c, \"gateway.antigravity_models\", zap.String(\"protocol\", string(service.UniversalProtocolAntigravity))).Warn(\"capability_discovery_failed\", zap.Error(err))"
