+++ scripts/sentinels/pricing-availability.json
@@ -7,12 +7,17 @@
       "must_contain": [
         "TkRecordFailureFromErr",
         "tkFilterModelIDs",
-        "tkAntigravityDefaultModels",
         "tkGeminiDefaultModelsList",
-        "TkEvalBodyGuard(",
-        "TkEnrichForbiddenMessage("
+        "TkEvalBodyGuard("
       ],
-      "rationale": "gateway_handler.go has TK injection points across multiple flows: (1) failure taps at the Anthropic /v1/messages + Gemini /v1/messages error paths (TkRecordFailureFromErr), (2) the CatalogPolicy model-list projection on the whitelist path (tkFilterModelIDs), (3) the shape-preserving antigravity model filter (tkAntigravityDefaultModels), (4) the gemini /v1/models fallback re-point to the CatalogPolicy projection (tkGeminiDefaultModelsList) so the list converges with /pricing instead of returning the raw geminicli.DefaultModels (which still carried advertised_dead gemini-2.0-flash) — an upstream merge could silently revert this one-line branch back to the raw list and un-converge gemini /v1/models, (5) the pre-flight body-size guard on Messages + CountTokens entries (TkEvalBodyGuard) that fail-fast oversize requests before forward — see edge:us1 production data where opus-4-7 upstream rejects body >= ~940 KB with raw 403; and (6) the actionable-message enrichment for upstream 403 in handleFailoverExhausted + handleFailoverExhaustedSimple (TkEnrichForbiddenMessage). An upstream merge touching the forward-failed error handlers, the Models/AntigravityModels methods, or the entry hooks could silently drop any of these. If TkRecordFailureFromErr is lost, model_not_found failures from Anthropic and Gemini platforms stop being classified; if TkEvalBodyGuard is lost, claude-cli clients lose the pre-flight rejection and again hit upstream 403; if TkEnrichForbiddenMessage is lost, the 403 response collapses back to 'Upstream access forbidden, please contact administrator' without body-size hint."
+      "rationale": "gateway_handler.go has TK injection points across multiple flows: (1) failure taps at the Anthropic /v1/messages + Gemini /v1/messages error paths (TkRecordFailureFromErr), (2) the CatalogPolicy model-list projection on the whitelist path (tkFilterModelIDs), (3) the gemini /v1/models fallback re-point to the CatalogPolicy projection (tkGeminiDefaultModelsList) so the list converges with /pricing instead of returning the raw geminicli.DefaultModels (which still carried advertised_dead gemini-2.0-flash) — an upstream merge could silently revert this one-line branch back to the raw list and un-converge gemini /v1/models, and (4) the pre-flight body-size guard on Messages + CountTokens entries (TkEvalBodyGuard) that fail-fast oversize requests before forward — see edge:us1 production data where opus-4-7 upstream rejects body >= ~940 KB with raw 403. Antigravity default-model projection and 403 enrichment live in TK companions (gateway_handler_tk_model_list.go / gateway_handler_tk_forward_error.go)."
+    },
+    {
+      "path": "backend/internal/handler/gateway_handler_tk_model_list.go",
+      "must_contain": [
+        "tkAntigravityDefaultModels"
+      ],
+      "rationale": "Shape-preserving antigravity model filter owner after TK companion extraction. An upstream merge that drops tkAntigravityDefaultModels un-converges /antigravity/models with CatalogPolicy."
     },
     {
       "path": "backend/internal/handler/gateway_handler_chat_completions.go",
@@ -64,9 +69,10 @@
       "path": "backend/internal/handler/gateway_handler_tk_forward_error.go",
       "must_contain": [
         "TkRecordFailureFromErr",
-        "UpstreamFailoverError"
+        "UpstreamFailoverError",
+        "TkEnrichForbiddenMessage("
       ],
-      "rationale": "This TK companion file is the single implementation of TkRecordFailureFromErr — the helper that extracts upstream HTTP status from UpstreamFailoverError and forwards it to the availability classifier. If this file is lost (e.g. an upstream merge that removes all untracked files), all 5 tap call sites compile against a missing function and the build fails. The file also contains the UpstreamFailoverError unwrap logic that makes model_not_found classification work correctly (R-004 fix from review-20260506)."
+      "rationale": "This TK companion file is the single implementation of TkRecordFailureFromErr — the helper that extracts upstream HTTP status from UpstreamFailoverError and forwards it to the availability classifier. If this file is lost (e.g. an upstream merge that removes all untracked files), all 5 tap call sites compile against a missing function and the build fails. The file also contains the UpstreamFailoverError unwrap logic that makes model_not_found classification work correctly (R-004 fix from review-20260506), plus the Messages-path TkEnrichForbiddenMessage call site for actionable 403 enrichment."
     },
     {
       "path": "backend/internal/handler/pricing_catalog_handler_tk.go",
